// src/app/generic/generic-actions/enums/actions-type.enums.ts
export enum EnumActionsType {
    actionNone = 0,
    actionNew = 1,
    actionEdit = 2,
    actionDelete = 3,
    actionView = 4,
    actionSave = 5,
    actionCancel = 6,
    actionSearch = 7,
    actionExport = 8,
    actionPrint = 9,
    actionClose = 10,
    actionList = 11,
    actionOpen = 12,
    actionApply = 13,
    actionReset = 14,
    actionAccept = 15,
    actionLogout = 16,
    actionLogin = 17,


    openConfig = 100,
    openEntities = 101,
    openPersons = 102,
    openArticles = 103,
    openUsers = 104,
    openHome = 105,
    openQuotes  = 106,
    openDocuments = 107,
    openInvoices = 108


}